Split handlers into v1 and v2
Organize endpoint implementations based on the API versions
for improved maintainability and clarity
