--- plugins/orchestrator-backend/src/service/handlers.ts
@@ -1,37 +1,0 @@
-import {
-  WorkflowOverviewListResult,
-  WorkflowOverviewListResultDTO,
-} from '@janus-idp/backstage-plugin-orchestrator-common';
-
-import { SonataFlowService } from './SonataFlowService';
-
-export async function getWorkflowOverviewV1(
-  sonataFlowService: SonataFlowService,
-): Promise<WorkflowOverviewListResult> {
-  const overviews = await sonataFlowService.fetchWorkflowOverviews();
-  if (!overviews) {
-    throw new Error("Couldn't fetch workflow overviews");
-  }
-  const result: WorkflowOverviewListResult = {
-    items: overviews,
-    limit: 0,
-    offset: 0,
-    totalCount: overviews?.length ?? 0,
-  };
-  return result;
-}
-
-export async function getWorkflowOverviewV2(
-  sonataFlowService: SonataFlowService,
-): Promise<WorkflowOverviewListResultDTO> {
-  const overviewsV1 = await getWorkflowOverviewV1(sonataFlowService);
-  const result: WorkflowOverviewListResultDTO = {
-    overviews: overviewsV1.items,
-    paginationInfo: {
-      limit: 0,
-      offset: 0,
-      totalCount: overviewsV1.items?.length ?? 0,
-    },
-  };
-  return result;
-}
